Profile update
User can now modify his profile, and by doing so change events that will be displayed. Minor bug fixes and ui changes.
There is a bug that needs to be fixed, as it stands at the moment user must change every option or else the unchanged options will go the their default values. Will fix it tomorrow im so tired...